+++ sisana/preprocessing/preprocessing.py
@@ -51,6 +51,10 @@ def preprocess_data(exp: str, datatype: str, number: int, outdir: str):
     cutdf = cutdf.drop(columns=['num_samps_expressed'])
 
     cutdf.columns = expdf.columns[:-1]
+    nsamps_remaining = len(cutdf.columns)    
+        
+    with open("./tmp/num_samples.txt", "w") as f:
+        f.write(str(nsamps_remaining))
     
     # Print summary statistics for the filtering of exp files
     dist = expdf["num_samps_expressed"].value_counts()